dribble(print-dribble): in-package & dribble*: reproducibility.
It starts and finishes dribbling with every file loading, which is not
perfect because it messes up with the current dribbling session.

But then, if one loads the file while dribbling, it's back to initial
state anyway:

;; Started
(whatever)
(load dribble-file)
;; Finished
;; Started
(foo)
;; => x
(dribble*)
;; Finished

So no form is skipped inside dribble file. There might be a problem
with file contention in OS, but I dunno, seems unlikely?