break: Another round.
Well, now it's clear BREAK* has to be a macro. Calling BREAK
recursively makes it break inside the BREAK* itself. Chicken and egg
problem.

BREAK* might still fill a niche, because the idea is good: listing
function name and symbol values is an immediate and useful
benefit. Need to brainstorm it some more...

If it's a macro, then detecting the current function suddenly becomes
really easy: CCL even has a special variable just for that!